File Size Validation

+++ source/src/user/UserService.js
@@ -73,10 +73,12 @@ const getUser = async (id) => {
 const updateUser = async (id, updateBody) => {
   const user = await User.findOne({ where: { id: id } });
   user.username = updateBody.username;
-  if (user.image) {
-    await FileService.deleteProfileImage(user.image);
+  if (updateBody.image) {
+    if (user.image) {
+      await FileService.deleteProfileImage(user.image);
+    }
+    user.image = await FileService.saveProfileImage(updateBody.image);
   }
-  user.image = await FileService.saveProfileImage(updateBody.image);
   await user.save();
   return {
     id: id,
